+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v1.6.7 (2023-11-09)
+
+- add cents: "no", "ifAny" formatting options
+
+## v1.6.6 (2023-10-12)
+
+- add support for cents in string template inputs in V2
+- fix V2 import in examples
+
+## v1.6.5 (2023-10-08)
+
+- include V2 beta into build
+
 ## v2.0.0-beta.0 (2023-10-07)
 
 non-breaking